requirements.txt
cohere
fastapi
uvicorn
pydantic
pydantic[email]
pyjwt
simple_salesforce
semantic_router
python-dotenv
requests
itsdangerous